refactored repl into its own module
